+++ src/types/TStep.ts
@@ -3,7 +3,7 @@
 //
 //
 //
-import { JOIN_TYPE, REMOVE_DUPLICATES_METHOD, REMOVE_DUPLICATES_STRATEGY, SORT_ORDER } from "./DataTable"
+import { JOIN_TYPE, REMOVE_DUPLICATES_METHOD, REMOVE_DUPLICATES_STRATEGY, SORT_ORDER, TOrderBy } from "./DataTable"
 import { TJson } from "./TJson"
 import { TSchemaRequestDelete, TSchemaRequestInsert, TSchemaRequestListEntities, TSchemaRequestSelect, TSchemaRequestUpdate } from "./TSchemaRequest"
 
@@ -39,11 +39,9 @@ export type TStepJoin = {
     "right-field": string
 }
 
-export type TStepFields = string
+export type TStepFields = string //NOSONAR
 
-export type TStepSort = {
-    [field: string]: SORT_ORDER
-}
+export type TStepSort = TOrderBy
 
 export type TStepRun = {
     ai: string